% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_table}
\alias{display_table}
\title{Create an HTML Table to Display a Data Frame}
\usage{
display_table(df)
}
\arguments{
\item{df}{data frame to display}
}
\description{
Create an HTML Table to Display a Data Frame
}
