% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_day_of_year}
\alias{first_day_of_year}
\alias{last_day_of_year}
\title{Return the first/last day of a given year as a Date}
\usage{
first_day_of_year(y)

last_day_of_year(y)
}
\arguments{
\item{y}{the year for which to return the first/last day}
}
\description{
Return the first/last day of a given year as a Date
}
