% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{period_length}
\alias{period_length}
\title{Calculates the duration between two dates}
\usage{
period_length(begin, end, unit = "year")
}
\arguments{
\item{begin}{beginning date of the period}

\item{end}{end date of the period}

\item{unit}{unit of measure for the duration of the period.
Can be seconds, minutes, hours, days, weeks, months, or
years.}
}
\description{
Calculates the duration between two dates
}
