% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_date_in_year}
\alias{seq_date_in_year}
\title{Sample uniformly dates withing a given year}
\usage{
seq_date_in_year(y, n)
}
\arguments{
\item{y}{the year from which the dates are sampled}

\item{n}{the number of dates to sample in the year}
}
\value{
a vector of dates
}
\description{
Sample uniformly dates withing a given year
}
