% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations-losses.R
\name{simulate_loss_count}
\alias{simulate_loss_count}
\title{Simulate the Loss Count by Policies in a Portfolio}
\usage{
simulate_loss_count(portfolio, initial_freq, initial_freq_date, freq_trend)
}
\arguments{
\item{portfolio}{A data frame containing the policy data.
It needs to contain the following columns:
\itemize{
\item \code{inception_date} \if{html}{\out{<Date>}}
\item \code{expiration_date} \if{html}{\out{<Date>}}
\item \code{n_expo} \if{html}{\out{<numeric>}}
}}

\item{initial_freq}{The expected annual loss frequency per
exposure at date \code{initial_freq_date}}

\item{initial_freq_date}{The date for the initial frequency}

\item{freq_trend}{The annual rate of change of the frequency}
}
\description{
Given a portfolio of policies and frequency assumptions
this function will simulate the loss count of each policy.
}
