% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_portfolio}
\alias{simulate_portfolio}
\title{Simulates a portfolio of insurance policies}
\usage{
simulate_portfolio(
  sim_years,
  initial_policy_count,
  ptf_growth,
  n_expo_per_policy,
  policy_length,
  initial_avg_premium,
  premium_trend,
  rate_change_data
)
}
\arguments{
\item{sim_years}{A vector with the simulated years}

\item{initial_policy_count}{The number of policies for the first simulated
year}

\item{ptf_growth}{The annual rate of growth of the portfolio size (in
policy count)}

\item{n_expo_per_policy}{The number of exposure associated to each policy}

\item{policy_length}{The policy length in months}

\item{initial_avg_premium}{The average premium per exposure at the beginning
of the first simulated year.}

\item{premium_trend}{The annual rate of growth of the average premium}

\item{rate_change_data}{A data frame containing the rate change information,
on row for each historical rate change.
It must contain two columns: 1) \code{effective_date}: the \code{Date} at which the
rate change has taken plance. 2) \code{rate_change}: the percentage increase
(if positive) or decrease (if negative) of the average premium after the
rate change.}
}
\value{
A data frame with a row for each simulated policy and the following
columns:
\itemize{
\item \code{policy_id}: a unique policy identifier
\item \code{inception_date}: the effective date of the simulated policy
\item \code{expiration_date}: the expiration date
\item \code{n_expo}: the number of exposures of the policy
\item \code{premium}: the policy premium
}
}
\description{
Based on the inputs this functions will generate a table
of simulates policies.
}
